+++ packages/br-validators/src/index.ts
@@ -3,6 +3,7 @@ export * from './cpf.js';
 export * from './cep.js';
 export * from './telefone.js';
 export * from './cnh.js';
+export * from './renavam.js';
 export * from './brcode.js';
 export * from './placa.js';
 export * from './pis-pasep.js';