+++ packages/br-validators/src/strip/index.ts
@@ -2,4 +2,5 @@ export { stripCep } from './cep.js';
 export { stripCpf } from './cpf.js';
 export { stripCnpj } from './cnpj.js';
 export { stripCnh } from './cnh.js';
+export { stripRenavam } from './renavam.js';
 export { stripPlaca } from './placa.js';